obsidian-kms-adk/requirements.txt
# Google ADK and AI Framework
google-adk>=0.1.0
google-generativeai>=0.8.0

# Web Framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
websockets>=12.0

# Database and Storage
neo4j>=5.15.0
psycopg2-binary>=2.9.7
pgvector>=0.2.4
redis>=5.0.1
sqlalchemy>=2.0.0

# Vector and Embeddings
sentence-transformers>=2.2.2
numpy>=1.24.0
scikit-learn>=1.3.0

# Markdown and Text Processing
python-frontmatter>=1.1.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
nltk>=3.8.1
spacy>=3.7.0

# Graph Processing
networkx>=3.2.0
pyvis>=0.3.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0
aiofiles>=23.0.0
watchdog>=3.0.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Monitoring and Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
